Prevent LAME-related buffer overruns in writer (TBC)

+++ dsp++/test/test_utility.h
@@ -37,7 +37,7 @@ struct temp_file {
 inline size_t copy_file(snd::reader& r, snd::writer& w) {
     const size_t bufsize = 512;
     size_t total = 0;
-    std::unique_ptr<float[]> buf{new float[r.format().channel_count() * bufsize]};
+    std::unique_ptr<short[]> buf{new short[r.format().channel_count() * bufsize]};
     while (true) {
         size_t read = r.read_frames(&buf[0], bufsize);
         total += w.write_frames(&buf[0], read);
